Backend-Register-Login

+++ django-backend/src/posts/api.py
@@ -1,7 +1,8 @@
 from ninja import Router
 from .models import Post
-from .schemas import PostInputSchema, PostOutputSchema
+from .schemas import PostInputSchema, PostOutputSchema, ErrorOutputSchema
 from django.shortcuts import get_object_or_404
+from ninja_jwt.authentication import JWTAuth
 
 # Endpoints here would route to the main api instance defined in src/config/api.py
 # so instead of using @api.get/post/put/delete use @router.get/post/put/delete instead
@@ -9,46 +10,67 @@ from django.shortcuts import get_object_or_404
 # You can test endpoints at http://127.0.0.1:8000/api/docs  (Make sure server is running)
 
 router = Router()
-
-# Response status codes help us see if the endpoint requests were successful or not.
-
-
+    
 # ENDPOINTS
+# Protected Endpoints require JWT authentication (Login)
 
-# Create a post
-@router.post("/", response={201:PostOutputSchema}) # response status 201 Created
-def create_post(request, data: PostInputSchema):
-    post = Post(
-        title=data.title,
-        content=data.content,
-    )
-    post.save()
-    return 201, post
+# To test and use Protected Endpoints: 
+# 1. You must successfully execute the login POST endpoint (Register a user first if you haven't already)
+# 2. After logging in, you'll be given a response body with an access token
+# 3. On the top right of the http://127.0.0.1:8000/api/docs/api/docs page, click on Authorize
+# 4. Copy and paste the access token (without the double quotes) and click Authorize
+# 5. Now you are now authorized to use Protected Endpoints.
 
-# List all posts
-@router.get("/", response=list[PostOutputSchema]) # response status is 200 OK by default
-def list_posts(request):
-    return Post.objects.all()
 
-# Get a single post by its id 
-@router.get("/{post_id}", response=PostOutputSchema) # response status is 200 OK by default
+# Get all posts 
+@router.get("/", response={200: list[PostOutputSchema]})
+def get_posts(request):
+    posts = Post.objects.order_by("-created_at")
+    return 200, [PostOutputSchema.from_orm(post) for post in posts]
+
+# Get all posts created by the user (PROTECTED)
+@router.get("/my-posts", auth=JWTAuth(), response={200: list[PostOutputSchema]})
+def get_my_posts(request):
+    posts = Post.objects.filter(author=request.user).order_by("-created_at")
+    return 200, [PostOutputSchema.from_orm(post) for post in posts]
+
+# Get a single post by its id
+@router.get("/{post_id}", response={200: PostOutputSchema}) 
 def get_post(request, post_id: int):
     post = get_object_or_404(Post, id=post_id)
-    return post
+    return 200, PostOutputSchema.from_orm(post)
 
-@router.put("/{post_id}", response=PostOutputSchema) # response status is 200 OK by default
+# Create a post (PROTECTED)
+@router.post("/", auth=JWTAuth(), response={201: PostOutputSchema}) 
+def create_post(request, data: PostInputSchema):
+    post = Post.objects.create(
+        title=data.title,
+        content=data.content,
+        author=request.user
+    )
+    return 201, PostOutputSchema.from_orm(post)
+
+# Update a single post by its id (PROTECTED)
+@router.put("/{post_id}", auth=JWTAuth(), response={200: PostOutputSchema, 403: ErrorOutputSchema})
 def update_post(request, post_id: int, data: PostInputSchema):
     post = get_object_or_404(Post, id=post_id)
+
+    if post.author != request.user:
+         return 403, {"error": "You do not have permission to modify this post"}
 
     post.title = data.title
     post.content = data.content
 
     post.save()
-    return post
+    return 200, PostOutputSchema.from_orm(post)
 
-# Delete a post by its id
-@router.delete("/{post_id}", response={204: None}) # response status is 204 No Content
+# Delete a single post by its id (PROTECTED)
+@router.delete("/{post_id}", auth=JWTAuth(), response={204: None, 403: ErrorOutputSchema}) 
 def delete_post(request, post_id: int):
     post = get_object_or_404(Post, id=post_id)
+
+    if post.author != request.user:
+        return 403, {"error": "You do not have permission to delete this post."}
+    
     post.delete()
-    return 204, None  
+    return 204, None
